fix(qc-research,#12834): porter les tests dedup dans test_qc_research_monitor canonique
Le test ajoute a scripts/tests/test_qc_research_monitor.py importait
notebook_tools.qc_research_monitor via sys.path sur scripts/ -- hors
convention du module (tests dans scripts/notebook_tools/tests/, import
direct from qc_research_monitor). En CI, notebooks_tools n'y est pas un
package (pas de __init__.py) -> ImportError Scripts Tests (CPU).

Fusion des 5 tests de dedup (reprise run 32690576863, suffixe "(#<article_id>)")
dans le fichier canonique, suppression du doublon. 15/15 tests passent.
